l6t2 форматирование

+++ src/lesson3/task1/Loop.kt
@@ -160,7 +160,8 @@ fun isCoPrime(m: Int, n: Int): Boolean {
  * то есть, существует ли такое целое k, что m <= k*k <= n.
  * Например, для интервала 21..28 21 <= 5*5 <= 28, а для интервала 51..61 квадрата не существует.
  */
-fun squareBetweenExists(m: Int, n: Int): Boolean = ceil(sqrt(m.toDouble())).toInt() <= floor(sqrt(n.toDouble())).toInt()
+fun squareBetweenExists(m: Int, n: Int): Boolean =
+        ceil(sqrt(m.toDouble())) <= floor(sqrt(n.toDouble()))
 
 
 /**
